requirements.txt
flask
flask-login
google-auth
google-auth-oauthlib
python-dotenv
requests
openai
langchain-openai
langchain-core
langchain
psycopg2
